Task2_Real_Time_Streaming/requirements.txt
kafka-python==2.0.2
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
